# sdk/openapi/aisdlc-integration-v1.yaml
openapi: 3.1.0
info:
  title: AI-SDLC Integration API
  version: 1.0.0
  description: |
    Versioned public integration surface for AI-SDLC developer tooling. The management server
    remains the authoritative OpenAPI publisher at `/v3/api-docs`; this checked-in contract is the
    compatibility baseline used to generate SDK artifacts.
servers:
  - url: https://control.example.invalid
security:
  - bearerAuth: []
paths:
  /api/v1/projects/{projectId}/scm-events:
    get:
      operationId: listScmEvents
      tags: [SCM]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Project-scoped SCM event ledger.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageScmEvent'
  /api/v1/projects/{projectId}/risk-intelligence/latest:
    get:
      operationId: getLatestRiskScore
      tags: [Risk]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Latest explainable risk snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskScore'
  /api/v1/projects/{projectId}/notification-channels:
    get:
      operationId: listNotificationChannels
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Safe notification channel metadata.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationChannel'
    post:
      operationId: createNotificationChannel
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationChannel'
      responses:
        '201':
          description: Channel created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identifier'
  /api/v1/projects/{projectId}/approvals:
    get:
      operationId: listApprovals
      tags: [Approvals]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Project approval queue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageApproval'
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Identifier:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
    NotificationChannel:
      type: object
      required: [id, type, name, enabled, destinationFingerprint]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [EMAIL, SLACK_WEBHOOK, TEAMS_WEBHOOK, GENERIC_WEBHOOK] }
        name: { type: string }
        enabled: { type: boolean }
        destinationFingerprint: { type: string }
    CreateNotificationChannel:
      type: object
      required: [type, name, destination]
      properties:
        type: { type: string, enum: [EMAIL, SLACK_WEBHOOK, TEAMS_WEBHOOK, GENERIC_WEBHOOK] }
        name: { type: string, minLength: 1, maxLength: 160 }
        destination: { type: string, maxLength: 2000 }
        sharedSecret: { type: string, writeOnly: true, maxLength: 2048 }
    ScmEvent:
      type: object
      required: [id, eventType, deliveryId, receivedAt]
      properties:
        id: { type: string, format: uuid }
        eventType: { type: string }
        deliveryId: { type: string }
        receivedAt: { type: string, format: date-time }
    RiskScore:
      type: object
      required: [id, score, band, formulaVersion, computedAt]
      properties:
        id: { type: string, format: uuid }
        score: { type: integer, minimum: 0, maximum: 100 }
        band: { type: string, enum: [LOW, MODERATE, HIGH, CRITICAL] }
        formulaVersion: { type: string }
        computedAt: { type: string, format: date-time }
        components: { type: object, additionalProperties: { type: number } }
    Approval:
      type: object
      required: [id, title, status, quorum, dueAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { type: string, enum: [PENDING, APPROVED, REJECTED, ESCALATED] }
        quorum: { type: integer, minimum: 1 }
        dueAt: { type: string, format: date-time }
    PageScmEvent:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ScmEvent' }
    PageApproval:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/Approval' }
    PageMetadata:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content: { type: array }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }